LayoutTests/accessibility/aria-description-expected.txt

This test ensures that aria-description maps to appropriate attributes and works with aria-describedby correctly.

Help text:
Help text: AXHelp: text
Help text: AXHelp: text

PASS successfullyParsed is true

TEST COMPLETE
button
button
text
button
text
